+++ syscall/samples/ioctl/BLKGETSIZE.c
@@ -5,13 +5,20 @@
 #include <unistd.h>
 #include <linux/fs.h>
 
-int main(void)
+int main(int argc, char *argv[])
 {
 	int err, fd;
 	unsigned long long range[2] = {0, 0};
-	char *dev = "/dev/sda";
+	char *dev;
 
-	fd = open(dev, O_RDONLY);
+	if (argc < 2) {
+		fprintf(stderr, "Usage: %s [dev]\n", argv[0]);
+		exit(1);
+	}
+
+	dev = argv[1];
+
+	fd = open(dev, O_RDWR);
 	if (fd == -1) {
 		fprintf(stderr, "open(%s) %m\n", dev);
 		exit(1);
@@ -22,10 +29,15 @@ int main(void)
 	 */
 	err = ioctl(fd, BLKGETSIZE, &range[1]);
 	if (err == -1) {
-		fprintf(stderr, "ioctl(): %m\n");
+		fprintf(stderr, "ioctl(%s, BLKGETSIZE): %m\n", dev);
 	}
 
 	printf("device %s size %lld GB\n", dev, range[1] * 512 / 1024 / 1024 / 1024);
+
+	err = ioctl(fd, BLKZEROOUT, range);
+	if (err == -1) {
+		fprintf(stderr, "ioctl(%s, BLKZEROOUT): %m\n", dev);
+	}
 
 	close(fd);
 	return 0;